refactor: modularize integration testing
